requirements.txt
boxmot
opencv-python
numpy
ultralytics
python-dotenv
pygame
pyserial